halved hitstun on landing
