+++ fighters/common/src/general_statuses/damagefall.rs
@@ -10,42 +10,9 @@ pub fn install() {
 fn nro_hook(info: &skyline::nro::NroInfo) {
     if info.name == "common" {
         skyline::install_hooks!(
-            status_DamageFall_Main,
-            FighterStatusUniqProcessDamage_leave_stop_hook,
+            status_DamageFall_Main
         );
     }
-}
-
-#[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_FighterStatusUniqProcessDamage_leave_stop)]
-pub unsafe fn FighterStatusUniqProcessDamage_leave_stop_hook(fighter: &mut L2CFighterCommon, arg2: L2CValue, arg3: L2CValue) -> L2CValue {
-    println!("leave stop");
-    original!()(fighter, arg2, arg3);
-    let hashmap = fighter.local_func__fighter_status_damage_2();
-    let sdi_mul = hashmap["stop_delay_"].get_f32();
-    // get stick x/y length
-    let stick_x = ControlModule::get_stick_x(fighter.module_accessor);
-    let stick_y = ControlModule::get_stick_y(fighter.module_accessor);
-    // get base asdi distance
-    let base_asdi = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("hit_stop_delay_auto_mul"));
-    // mul sdi_mul by hit_stop_delay_auto_mul = total sdi
-    let asdi = sdi_mul * base_asdi;
-    // mul stick x/y by total sdi
-    let asdi_x = asdi * stick_x;
-    let asdi_y = asdi * stick_y;
-    // get current pos
-    let mut pos = Vector3f {
-        x: PostureModule::pos_x(fighter.module_accessor),
-        y: PostureModule::pos_y(fighter.module_accessor),
-        z: PostureModule::pos_z(fighter.module_accessor)
-    };
-    println!("leave stop pos: {}, {}, {}", pos.x, pos.y, pos.z);
-    // add asdi x/y to pos
-    pos.x += asdi_x;
-    pos.y += asdi_y;
-    PostureModule::set_pos(fighter.module_accessor, &Vector3f{x: pos.x, y: pos.y, z: pos.z});
-    // make sure we can enter tech/missed tech on f1 of damage fly statuses (vanilla only allows them starting on f3)
-    WorkModule::on_flag(fighter.module_accessor, *FIGHTER_STATUS_DAMAGE_FLAG_ENABLE_DOWN);
-    return 0.into()
 }
 
 #[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_status_DamageFall_Main)]
