+++ fighters/common/src/general_statuses/mod.rs
@@ -26,6 +26,7 @@ mod downdamage;
 mod crawl;
 mod cliff;
 mod catchcut;
+mod damage;
 // [LUA-REPLACE-REBASE]
 // [SHOULD-CHANGE]
 // Reimplement the whole status script (already done) instead of doing this.
@@ -148,7 +149,7 @@ fn nro_hook(info: &skyline::nro::NroInfo) {
             damage_fly_common_init, 
             //damage_air_main,
             status_Landing_MainSub,
-            //status_pre_Landing,
+            status_LandingStiffness,
             status_pre_LandingLight,
             status_LandingAttackAirSub,
             status_pre_landing_fall_special,
@@ -162,21 +163,13 @@ fn nro_hook(info: &skyline::nro::NroInfo) {
     }
 }
 
-#[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_status_pre_Landing)]
-pub unsafe fn status_pre_Landing(fighter: &mut L2CFighterCommon) -> L2CValue {
-    let id = VarModule::get_int(fighter.battle_object, vars::common::instance::COSTUME_SLOT_NUMBER) as usize;
-    let mut fighter_pos = Vector3f {
-        x: PostureModule::pos_x(fighter.module_accessor),
-        y: PostureModule::pos_y(fighter.module_accessor),
-        z: PostureModule::pos_z(fighter.module_accessor)
-    };
-    fighter_pos.y += VarModule::get_float(fighter.object(), vars::common::instance::ECB_Y_OFFSETS);
-    VarModule::set_float(fighter.battle_object, vars::common::instance::GET_DIST_TO_FLOOR, GroundModule::get_distance_to_floor(fighter.module_accessor, &fighter_pos, fighter_pos.y, true));
-    let dist = VarModule::get_float(fighter.battle_object, vars::common::instance::GET_DIST_TO_FLOOR);
-    if (0.0 <= dist || VarModule::is_flag(fighter.battle_object, vars::common::instance::ENABLE_AIR_ESCAPE_MAGNET)) && dist < 0.1 {
-        if dist != -1.0 {
-            PostureModule::set_pos(fighter.module_accessor, &fighter_pos);
-        }
+#[skyline::hook(replace = smash::lua2cpp::L2CFighterCommon_status_LandingStiffness)]
+pub unsafe fn status_LandingStiffness(fighter: &mut L2CFighterCommon) -> L2CValue {
+    if fighter.global_table[PREV_STATUS_KIND] == FIGHTER_STATUS_KIND_DAMAGE_AIR
+    || fighter.global_table[PREV_STATUS_KIND] == FIGHTER_STATUS_KIND_SAVING_DAMAGE_AIR {
+        // halve hitstun on landing
+        let hitstun = WorkModule::get_float(fighter.module_accessor, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_REACTION_FRAME);
+        WorkModule::set_float(fighter.module_accessor, hitstun * 0.5, *FIGHTER_INSTANCE_WORK_ID_FLOAT_DAMAGE_REACTION_FRAME);
     }
     original!()(fighter)
 }
@@ -334,6 +327,7 @@ pub fn install() {
     crawl::install();
     cliff::install();
     catchcut::install();
+    damage::install();
 
     smashline::install_status_scripts!(
         damage_fly_end,
